+++ src/TrafficParsing/SOCKS5/SOCKS5Parser.hpp
@@ -13,7 +13,7 @@ namespace Proxy::TrafficParsing
         static uint8_t GetClientAuthenticationMethod(const uint8_t* buffer, uint32_t bufferSize) noexcept;
         static uint8_t GetDestinationAddressType(const uint8_t* buffer, uint32_t bufferSize) noexcept;
         static Status GetDestinationAddressAndPort(const uint8_t* buffer, uint32_t bufferSize, ByteStream& destinationAddress, uint16_t& port) noexcept;
-        static size_t GetConnectionRequestLength(const uint8_t* buffer, size_t buffersize);
+        static int32_t GetConnectionRequestLength(const uint8_t* buffer, size_t buffersize);
 
     private:
         static void MemcpyDestinationAddressAndPortFromBuffer(const uint8_t* buffer, uint8_t destinationAddressSize, ByteStream& stream, uint16_t& port, uint8_t offset) noexcept;
